Preliminary gui for shortcuts, no real functionality yet, and some bug fixes (of course :-P)
git-svn-id: svn://localhost/tmp/konfetka/trunk@152 aaffa179-fd03-4778-91ff-b08a3235a277

+++ src/medialib.cpp
@@ -621,7 +621,7 @@ QMimeData* MediaLib::getCurrentMimeData() {
 // 	std::cout<<tmp<<std::endl;
 // 	stream << tmp;	
 // 	encodedData.append((char*)tmp);
-	encodedData =QByteArray::number((int)tmp);
+	encodedData =QByteArray::number((qlonglong)tmp);
 // 	std::cout<<tmp<<std::endl;
 
 	mimeData->setData("application/x-collstruct", encodedData);
